requirements.txt
pycryptodome
cryptography
pytest
pytest-cov
bandit
